+++ lcheapo/lccut.py
@@ -84,13 +84,14 @@ def main():
                         start_block += MAX_BLOCK_READ
     # Save/append process information to process_steps file
     sdpchain.make_process_steps_file(
+        args.in_dir,
+        args.out_dir,
         'lccut',
         __doc__,
         __version__,
         " ".join(sys.argv),
         start_time_str,
         return_code,
-        args.base_dir,
         exec_messages=exec_messages,
         exec_parameters='')